requirements.txt
PyQt5>=5.15.0
paho-mqtt>=1.6.1
python-dotenv>=0.19.0
pyqtgraph>=0.13.3
numpy>=1.26.0
setuptools>=68.0.0
